Move around files and imports
